+++ tests/Unit/Helper/Filter/JSON/FixRecipeYieldFilterTest.php
@@ -4,7 +4,7 @@ namespace OCA\Cookbook\tests\Unit\Helper\Filter\JSON;
 
 use OCA\Cookbook\Helper\Filter\JSON\FixRecipeYieldFilter;
 use OCP\IL10N;
-use OCP\ILogger;
+use Psr\Log\LoggerInterface;
 use PHPUnit\Framework\MockObject\Stub;
 use PHPUnit\Framework\TestCase;
 
@@ -20,8 +20,8 @@ class FixRecipeYieldFilterTest extends TestCase {
 		$l = $this->createStub(IL10N::class);
 		$l->method('t')->willReturnArgument(0);
 
-		/** @var ILogger */
-		$logger = $this->createStub(ILogger::class);
+		/** @var LoggerInterface */
+		$logger = $this->createStub(LoggerInterface::class);
 
 		$this->dut = new FixRecipeYieldFilter($l, $logger);